backend/requirements.txt
fastapi
uvicorn[standard]
langchain
faiss-cpu
sentence-transformers
python-dotenv
pydantic>=2.0
langchain-community
